Move localStorage to a separate file
