+++ client/src/components/Form/Form.js
@@ -2,20 +2,9 @@ import React, { useState } from 'react';
 
 import useDropdown from '../../hooks/useDropdown';
 import fetchData from '../../utils/fetchData';
+import useStateWithLocalStorage from '../../utils/localStorage';
 
 import Search from '../Search/Search';
-
-const useStateWithLocalStorage = localStorageKey => {
-  const [query, setQuery] = React.useState(
-    localStorage.getItem(localStorageKey) || ''
-  );
-
-  React.useEffect(() => {
-    localStorage.setItem(localStorageKey, query);
-  }, [query]);
-
-  return [query, setQuery];
-};
 
 const Form = ({ setData, setIsError, setIsLoading, navigate }) => {
   const options = ['login', 'extensions/user_id'];
