+++ client/src/utils/localStorage.js
@@ -1,0 +1,15 @@
+import React, { useEffect } from 'react';
+
+const useStateWithLocalStorage = localStorageKey => {
+  const [query, setQuery] = React.useState(
+    localStorage.getItem(localStorageKey) || ''
+  );
+
+  useEffect(() => {
+    localStorage.setItem(localStorageKey, query);
+  }, [query]);
+
+  return [query, setQuery];
+};
+
+export { useStateWithLocalStorage };
